+++ src/communication.h
@@ -6,7 +6,7 @@
 void initCommunication();
 void addDataToBuffer(uint16_t adc_data, uint32_t afe_data);
 void sendBufferIfFull();
-void addDataToBufferSingle(uint16_t adc_data);
+void addDataToBufferSingle(uint32_t adc_data);
 void sendBufferIfFullSingle();
 
 #endif // COMMUNICATION_H
